requirements.txt
langgraph
langchain
langchain-openai
langchain-core
opik
python-dotenv
pydantic